+++ examples/simple-example.rs
@@ -4,7 +4,7 @@ use std::marker::PhantomData;
 
 use halo2::{
     arithmetic::FieldExt,
-    circuit::{Cell, Chip, Layouter, Region, SimpleFloorPlanner},
+    circuit::{AssignedCell, Chip, Layouter, Region, SimpleFloorPlanner},
     plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Fixed, Instance, Selector},
     poly::Rotation,
 };
@@ -65,9 +65,6 @@ struct FieldConfig {
     // This is important when building larger circuits, where columns are used by
     // multiple sets of instructions.
     s_mul: Selector,
-
-    /// The fixed column used to load constants.
-    constant: Column<Fixed>,
 }
 
 impl<F: FieldExt> FieldChip<F> {
@@ -84,10 +81,10 @@ impl<F: FieldExt> FieldChip<F> {
         instance: Column<Instance>,
         constant: Column<Fixed>,
     ) -> <Self as Chip<F>>::Config {
-        meta.enable_equality(instance.into());
+        meta.enable_equality(instance);
         meta.enable_constant(constant);
         for column in &advice {
-            meta.enable_equality((*column).into());
+            meta.enable_equality(*column);
         }
         let s_mul = meta.selector();
 
@@ -125,7 +122,6 @@ impl<F: FieldExt> FieldChip<F> {
             advice,
             instance,
             s_mul,
-            constant,
         }
     }
 }
@@ -149,10 +145,7 @@ impl<F: FieldExt> Chip<F> for FieldChip<F> {
 // ANCHOR: instructions-impl
 /// A variable representing a number.
 #[derive(Clone)]
-struct Number<F: FieldExt> {
-    cell: Cell,
-    value: Option<F>,
-}
+struct Number<F: FieldExt>(AssignedCell<F, F>);
 
 impl<F: FieldExt> NumericInstructions<F> for FieldChip<F> {
     type Num = Number<F>;
@@ -164,21 +157,19 @@ impl<F: FieldExt> NumericInstructions<F> for FieldChip<F> {
     ) -> Result<Self::Num, Error> {
         let config = self.config();
 
-        let mut num = None;
         layouter.assign_region(
             || "load private",
             |mut region| {
-                let cell = region.assign_advice(
-                    || "private input",
-                    config.advice[0],
-                    0,
-                    || value.ok_or(Error::SynthesisError),
-                )?;
-                num = Some(Number { cell, value });
-                Ok(())
+                region
+                    .assign_advice(
+                        || "private input",
+                        config.advice[0],
+                        0,
+                        || value.ok_or(Error::Synthesis),
+                    )
+                    .map(Number)
             },
-        )?;
-        Ok(num.unwrap())
+        )
     }
 
     fn load_constant(
@@ -188,24 +179,14 @@ impl<F: FieldExt> NumericInstructions<F> for FieldChip<F> {
     ) -> Result<Self::Num, Error> {
         let config = self.config();
 
-        let mut num = None;
         layouter.assign_region(
             || "load constant",
             |mut region| {
-                let cell = region.assign_advice_from_constant(
-                    || "constant value",
-                    config.advice[0],
-                    0,
-                    constant,
-                )?;
-                num = Some(Number {
-                    cell,
-                    value: Some(constant),
-                });
-                Ok(())
+                region
+                    .assign_advice_from_constant(|| "constant value", config.advice[0], 0, constant)
+                    .map(Number)
             },
-        )?;
-        Ok(num.unwrap())
+        )
     }
 
     fn mul(
@@ -216,7 +197,6 @@ impl<F: FieldExt> NumericInstructions<F> for FieldChip<F> {
     ) -> Result<Self::Num, Error> {
         let config = self.config();
 
-        let mut out = None;
         layouter.assign_region(
             || "mul",
             |mut region: Region<'_, F>| {
@@ -229,38 +209,25 @@ impl<F: FieldExt> NumericInstructions<F> for FieldChip<F> {
                 // but we can only rely on relative offsets inside this region. So we
                 // assign new cells inside the region and constrain them to have the
                 // same values as the inputs.
-                let lhs = region.assign_advice(
-                    || "lhs",
-                    config.advice[0],
-                    0,
-                    || a.value.ok_or(Error::SynthesisError),
-                )?;
-                let rhs = region.assign_advice(
-                    || "rhs",
-                    config.advice[1],
-                    0,
-                    || b.value.ok_or(Error::SynthesisError),
-                )?;
-                region.constrain_equal(a.cell, lhs)?;
-                region.constrain_equal(b.cell, rhs)?;
+                a.0.copy_advice(|| "lhs", &mut region, config.advice[0], 0)?;
+                b.0.copy_advice(|| "rhs", &mut region, config.advice[1], 0)?;
 
-                // Now we can assign the multiplication result into the output position.
-                let value = a.value.and_then(|a| b.value.map(|b| a * b));
-                let cell = region.assign_advice(
-                    || "lhs * rhs",
-                    config.advice[0],
-                    1,
-                    || value.ok_or(Error::SynthesisError),
-                )?;
+                // Now we can assign the multiplication result, which is to be assigned
+                // into the output position.
+                let value = a.0.value().and_then(|a| b.0.value().map(|b| *a * *b));
 
-                // Finally, we return a variable representing the output,
-                // to be used in another part of the circuit.
-                out = Some(Number { cell, value });
-                Ok(())
+                // Finally, we do the assignment to the output, returning a
+                // variable to be used in another part of the circuit.
+                region
+                    .assign_advice(
+                        || "lhs * rhs",
+                        config.advice[0],
+                        1,
+                        || value.ok_or(Error::Synthesis),
+                    )
+                    .map(Number)
             },
-        )?;
-
-        Ok(out.unwrap())
+        )
     }
 
     fn expose_public(
@@ -271,7 +238,7 @@ impl<F: FieldExt> NumericInstructions<F> for FieldChip<F> {
     ) -> Result<(), Error> {
         let config = self.config();
 
-        layouter.constrain_instance(num.cell, config.instance, row)
+        layouter.constrain_instance(num.0.cell(), config.instance, row)
     }
 }
 // ANCHOR_END: instructions-impl