+++ src/dev.rs
@@ -1,6 +1,7 @@
 //! Tools for developing circuits.
 
 use std::collections::HashMap;
+use std::collections::HashSet;
 use std::fmt;
 use std::iter;
 use std::ops::{Add, Mul, Neg, Range};
@@ -12,12 +13,13 @@ use crate::{
     arithmetic::{FieldExt, Group},
     plonk::{
         permutation, Advice, Any, Assignment, Circuit, Column, ColumnType, ConstraintSystem, Error,
-        Expression, Fixed, FloorPlanner, Instance, Selector,
+        Expression, Fixed, FloorPlanner, Instance, Selector, VirtualCell,
     },
     poly::Rotation,
 };
 
 pub mod metadata;
+mod util;
 
 pub mod cost;
 pub use cost::CircuitCost;
@@ -32,11 +34,34 @@ mod graph;
 #[cfg_attr(docsrs, doc(cfg(feature = "dev-graph")))]
 pub use graph::{circuit_dot_graph, layout::CircuitLayout};
 
+/// The location at which a particular lookup is not satisfied.
+#[derive(Debug, PartialEq)]
+pub enum LookupFailure {
+    /// A location inside a region. This may be due to the intentional use of a lookup
+    /// (if its inputs are conditional on a complex selector), or an unintentional lookup
+    /// constraint that overlaps the region (indicating that the lookup's inputs should be
+    /// made conditional).
+    InRegion {
+        /// The region in which the lookup's input expressions were evaluated.
+        region: metadata::Region,
+        /// The offset (relative to the start of the region) at which the lookup's input
+        /// expressions were evaluated.
+        offset: usize,
+    },
+    /// A location outside of a region. This most likely means the input expressions do
+    /// not correctly constrain a default value that exists in the table when the lookup
+    /// is not being used.
+    OutsideRegion {
+        /// The circuit row on which this lookup is not satisfied.
+        row: usize,
+    },
+}
+
 /// The reasons why a particular circuit is not satisfied.
 #[derive(Debug, PartialEq)]
 pub enum VerifyFailure {
     /// A cell used in an active gate was not assigned to.
-    Cell {
+    CellNotAssigned {
         /// The index of the active gate.
         gate: metadata::Gate,
         /// The region in which this cell should be assigned.
@@ -54,6 +79,8 @@ pub enum VerifyFailure {
         constraint: metadata::Constraint,
         /// The row on which this constraint is not satisfied.
         row: usize,
+        /// The values of the virtual cells used by this constraint.
+        cell_values: Vec<(metadata::VirtualCell, String)>,
     },
     /// A constraint was active on an unusable row, and is likely missing a selector.
     ConstraintPoisoned {
@@ -66,8 +93,8 @@ pub enum VerifyFailure {
         /// the order in which `ConstraintSystem::lookup` is called during
         /// `Circuit::configure`.
         lookup_index: usize,
-        /// The row on which this lookup is not satisfied.
-        row: usize,
+        /// The location at which the lookup is not satisfied.
+        location: LookupFailure,
     },
     /// A permutation did not preserve the original value of a cell.
     Permutation {
@@ -81,7 +108,7 @@ pub enum VerifyFailure {
 impl fmt::Display for VerifyFailure {
     fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
         match self {
-            Self::Cell {
+            Self::CellNotAssigned {
                 gate,
                 region,
                 column,
@@ -93,8 +120,16 @@ impl fmt::Display for VerifyFailure {
                     region, gate, column, offset
                 )
             }
-            Self::ConstraintNotSatisfied { constraint, row } => {
-                write!(f, "{} is not satisfied on row {}", constraint, row)
+            Self::ConstraintNotSatisfied {
+                constraint,
+                row,
+                cell_values,
+            } => {
+                writeln!(f, "{} is not satisfied on row {}", constraint, row)?;
+                for (name, value) in cell_values {
+                    writeln!(f, "- {} = {}", name, value)?;
+                }
+                Ok(())
             }
             Self::ConstraintPoisoned { constraint } => {
                 write!(
@@ -103,9 +138,19 @@ impl fmt::Display for VerifyFailure {
                     constraint
                 )
             }
-            Self::Lookup { lookup_index, row } => {
-                write!(f, "Lookup {} is not satisfied on row {}", lookup_index, row)
-            }
+            Self::Lookup {
+                lookup_index,
+                location,
+            } => match location {
+                LookupFailure::InRegion { region, offset } => write!(
+                    f,
+                    "Lookup {} is not satisfied in {} at offset {}",
+                    lookup_index, region, offset
+                ),
+                LookupFailure::OutsideRegion { row } => {
+                    write!(f, "Lookup {} is not satisfied on row {}", lookup_index, row)
+                }
+            },
             Self::Permutation { column, row } => {
                 write!(
                     f,
@@ -121,8 +166,10 @@ impl fmt::Display for VerifyFailure {
 struct Region {
     /// The name of the region. Not required to be unique.
     name: String,
-    /// The row that this region starts on, if known.
-    start: Option<usize>,
+    /// The columns involved in this region.
+    columns: HashSet<Column<Any>>,
+    /// The rows that this region starts and ends on, if known.
+    rows: Option<(usize, usize)>,
     /// The selectors that have been enabled in this region. All other selectors are by
     /// construction not enabled.
     enabled_selectors: HashMap<Selector, Vec<usize>>,
@@ -132,14 +179,20 @@ struct Region {
 }
 
 impl Region {
-    fn update_start(&mut self, row: usize) {
+    fn update_extent(&mut self, column: Column<Any>, row: usize) {
+        self.columns.insert(column);
+
         // The region start is the earliest row assigned to.
-        let mut start = self.start.unwrap_or(row);
+        // The region end is the latest row assigned to.
+        let (mut start, mut end) = self.rows.unwrap_or((row, row));
         if row < start {
             // The first row assigned was not at start 0 within the region.
             start = row;
         }
-        self.start = Some(start);
+        if row > end {
+            end = row;
+        }
+        self.rows = Some((start, end));
     }
 }
 
@@ -242,7 +295,7 @@ impl<F: Group + Field> Mul<F> for Value<F> {
 ///     circuit::{Layouter, SimpleFloorPlanner},
 ///     dev::{MockProver, VerifyFailure},
 ///     pasta::Fp,
-///     plonk::{Advice, Circuit, Column, ConstraintSystem, Error, Selector},
+///     plonk::{Advice, Any, Circuit, Column, ConstraintSystem, Error, Selector},
 ///     poly::Rotation,
 /// };
 /// const K: u32 = 5;
@@ -292,15 +345,15 @@ impl<F: Group + Field> Mul<F> for Value<F> {
 ///         layouter.assign_region(|| "Example region", |mut region| {
 ///             config.s.enable(&mut region, 0)?;
 ///             region.assign_advice(|| "a", config.a, 0, || {
-///                 self.a.map(|v| F::from(v)).ok_or(Error::SynthesisError)
+///                 self.a.map(|v| F::from(v)).ok_or(Error::Synthesis)
 ///             })?;
 ///             region.assign_advice(|| "b", config.b, 0, || {
-///                 self.b.map(|v| F::from(v)).ok_or(Error::SynthesisError)
+///                 self.b.map(|v| F::from(v)).ok_or(Error::Synthesis)
 ///             })?;
 ///             region.assign_advice(|| "c", config.c, 0, || {
 ///                 self.a
 ///                     .and_then(|a| self.b.map(|b| F::from(a * b)))
-///                     .ok_or(Error::SynthesisError)
+///                     .ok_or(Error::Synthesis)
 ///             })?;
 ///             Ok(())
 ///         })
@@ -321,12 +374,26 @@ impl<F: Group + Field> Mul<F> for Value<F> {
 ///     prover.verify(),
 ///     Err(vec![VerifyFailure::ConstraintNotSatisfied {
 ///         constraint: ((0, "R1CS constraint").into(), 0, "buggy R1CS").into(),
-///         row: 0
+///         row: 0,
+///         cell_values: vec![
+///             (((Any::Advice, 0).into(), 0).into(), "0x2".to_string()),
+///             (((Any::Advice, 1).into(), 0).into(), "0x4".to_string()),
+///             (((Any::Advice, 2).into(), 0).into(), "0x8".to_string()),
+///         ],
 ///     }])
 /// );
+///
+/// // If we provide a too-small K, we get an error.
+/// assert!(matches!(
+///     MockProver::<Fp>::run(2, &circuit, vec![]).unwrap_err(),
+///     Error::NotEnoughRowsAvailable {
+///         current_k,
+///     } if current_k == 2,
+/// ));
 /// ```
 #[derive(Debug)]
 pub struct MockProver<F: Group + Field> {
+    k: u32,
     n: u32,
     cs: ConstraintSystem<F>,
 
@@ -360,7 +427,8 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
         assert!(self.current_region.is_none());
         self.current_region = Some(Region {
             name: name().into(),
-            start: None,
+            columns: HashSet::default(),
+            rows: None,
             enabled_selectors: HashMap::default(),
             cells: vec![],
         });
@@ -376,7 +444,7 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
         AR: Into<String>,
     {
         if !self.usable_rows.contains(&row) {
-            return Err(Error::BoundsFailure);
+            return Err(Error::not_enough_rows_available(self.k));
         }
 
         // Track that this selector was enabled. We require that all selectors are enabled
@@ -396,7 +464,7 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
 
     fn query_instance(&self, column: Column<Instance>, row: usize) -> Result<Option<F>, Error> {
         if !self.usable_rows.contains(&row) {
-            return Err(Error::BoundsFailure);
+            return Err(Error::not_enough_rows_available(self.k));
         }
 
         self.instance
@@ -420,11 +488,11 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
         AR: Into<String>,
     {
         if !self.usable_rows.contains(&row) {
-            return Err(Error::BoundsFailure);
+            return Err(Error::not_enough_rows_available(self.k));
         }
 
         if let Some(region) = self.current_region.as_mut() {
-            region.update_start(row);
+            region.update_extent(column.into(), row);
             region.cells.push((column.into(), row));
         }
 
@@ -451,11 +519,11 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
         AR: Into<String>,
     {
         if !self.usable_rows.contains(&row) {
-            return Err(Error::BoundsFailure);
+            return Err(Error::not_enough_rows_available(self.k));
         }
 
         if let Some(region) = self.current_region.as_mut() {
-            region.update_start(row);
+            region.update_extent(column.into(), row);
             region.cells.push((column.into(), row));
         }
 
@@ -476,7 +544,7 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
         right_row: usize,
     ) -> Result<(), crate::plonk::Error> {
         if !self.usable_rows.contains(&left_row) || !self.usable_rows.contains(&right_row) {
-            return Err(Error::BoundsFailure);
+            return Err(Error::not_enough_rows_available(self.k));
         }
 
         self.permutation
@@ -490,7 +558,7 @@ impl<F: Field + Group> Assignment<F> for MockProver<F> {
         _: Option<Assigned<F>>,
     ) -> Result<(), Error> {
         if !self.usable_rows.contains(&from_row) {
-            return Err(Error::BoundsFailure);
+            return Err(Error::not_enough_rows_available(self.k));
         }
 
         Ok(())
@@ -524,11 +592,11 @@ impl<F: FieldExt> MockProver<F> {
         let cs = cs;
 
         if n < cs.minimum_rows() {
-            return Err(Error::NotEnoughRowsAvailable);
+            return Err(Error::not_enough_rows_available(k));
         }
 
         if instance.len() != cs.num_instance_columns {
-            return Err(Error::IncompatibleParams);
+            return Err(Error::InvalidInstances);
         }
 
         let instance = instance
@@ -564,6 +632,7 @@ impl<F: FieldExt> MockProver<F> {
         let constants = cs.constants.clone();
 
         let mut prover = MockProver {
+            k,
             n: n as u32,
             cs,
             regions: vec![],
@@ -626,11 +695,11 @@ impl<F: FieldExt> MockProver<F> {
                                 if r.cells.contains(&(cell.column, cell_row)) {
                                     None
                                 } else {
-                                    Some(VerifyFailure::Cell {
+                                    Some(VerifyFailure::CellNotAssigned {
                                         gate: (gate_index, gate.name()).into(),
                                         region: (r_i, r.name.clone()).into(),
                                         column: cell.column,
-                                        offset: cell_row as isize - r.start.unwrap() as isize,
+                                        offset: cell_row as isize - r.rows.unwrap().0 as isize,
                                     })
                                 }
                             })
@@ -697,6 +766,18 @@ impl<F: FieldExt> MockProver<F> {
                                     )
                                         .into(),
                                     row: (row - n) as usize,
+                                    cell_values: util::cell_values(
+                                        gate,
+                                        poly,
+                                        &load(n, row, &self.cs.fixed_queries, &self.fixed),
+                                        &load(n, row, &self.cs.advice_queries, &self.advice),
+                                        &load_instance(
+                                            n,
+                                            row,
+                                            &self.cs.instance_queries,
+                                            &self.instance,
+                                        ),
+                                    ),
                                 }),
                                 Value::Poison => Some(VerifyFailure::ConstraintPoisoned {
                                     constraint: (
@@ -777,9 +858,64 @@ impl<F: FieldExt> MockProver<F> {
                         if lookup_passes {
                             None
                         } else {
+                            let input_columns: HashSet<Column<Any>> = lookup
+                                .input_expressions
+                                .iter()
+                                .flat_map(|expression| {
+                                    expression.evaluate(
+                                        &|_| vec![],
+                                        &|_| {
+                                            panic!(
+                                                "virtual selectors are removed during optimization"
+                                            )
+                                        },
+                                        &|index, _, _| vec![self.cs.fixed_queries[index].0.into()],
+                                        &|index, _, _| vec![self.cs.advice_queries[index].0.into()],
+                                        &|index, _, _| {
+                                            vec![self.cs.instance_queries[index].0.into()]
+                                        },
+                                        &|a| a,
+                                        &|mut a, mut b| {
+                                            a.append(&mut b);
+                                            a
+                                        },
+                                        &|mut a, mut b| {
+                                            a.append(&mut b);
+                                            a
+                                        },
+                                        &|a, _| a,
+                                    )
+                                })
+                                .collect();
+                            println!("Lookup input columns: {:?}", &input_columns);
+
+                            // Figure out whether the lookup location
+                            // overlaps an assigned region.
+                            let location = self
+                                .regions
+                                .iter()
+                                .enumerate()
+                                .find(|(_, r)| {
+                                    let (start, end) = r.rows.unwrap();
+                                    // We match the region if any input columns overlap,
+                                    // rather than all of them, because matching complex
+                                    // selector columns is hard. This is probably going to
+                                    // cause some mismatches, which we'll fix later.
+                                    input_row >= start
+                                        && input_row <= end
+                                        && !input_columns.is_disjoint(&r.columns)
+                                })
+                                .map(|(r_i, r)| LookupFailure::InRegion {
+                                    region: (r_i, r.name.clone()).into(),
+                                    offset: input_row as usize - r.rows.unwrap().0 as usize,
+                                })
+                                .unwrap_or_else(|| LookupFailure::OutsideRegion {
+                                    row: input_row as usize,
+                                });
+
                             Some(VerifyFailure::Lookup {
                                 lookup_index,
-                                row: input_row as usize,
+                                location,
                             })
                         }
                     })
@@ -853,10 +989,10 @@ impl<F: FieldExt> MockProver<F> {
 mod tests {
     use pasta_curves::Fp;
 
-    use super::{MockProver, VerifyFailure};
+    use super::{LookupFailure, MockProver, VerifyFailure};
     use crate::{
         circuit::{Layouter, SimpleFloorPlanner},
-        plonk::{Advice, Any, Circuit, Column, ConstraintSystem, Error, Selector},
+        plonk::{Advice, Any, Circuit, Column, ConstraintSystem, Error, Selector, TableColumn},
         poly::Rotation,
     };
 
@@ -923,11 +1059,117 @@ mod tests {
         let prover = MockProver::run(K, &FaultyCircuit {}, vec![]).unwrap();
         assert_eq!(
             prover.verify(),
-            Err(vec![VerifyFailure::Cell {
+            Err(vec![VerifyFailure::CellNotAssigned {
                 gate: (0, "Equality check").into(),
                 region: (0, "Faulty synthesis".to_owned()).into(),
                 column: Column::new(1, Any::Advice),
                 offset: 1,
+            }])
+        );
+    }
+
+    #[test]
+    fn bad_lookup() {
+        const K: u32 = 4;
+
+        #[derive(Clone)]
+        struct FaultyCircuitConfig {
+            a: Column<Advice>,
+            q: Selector,
+            table: TableColumn,
+        }
+
+        struct FaultyCircuit {}
+
+        impl Circuit<Fp> for FaultyCircuit {
+            type Config = FaultyCircuitConfig;
+            type FloorPlanner = SimpleFloorPlanner;
+
+            fn configure(meta: &mut ConstraintSystem<Fp>) -> Self::Config {
+                let a = meta.advice_column();
+                let q = meta.complex_selector();
+                let table = meta.lookup_table_column();
+
+                meta.lookup(|cells| {
+                    let a = cells.query_advice(a, Rotation::cur());
+                    let q = cells.query_selector(q);
+
+                    // If q is enabled, a must be in the table.
+                    // Zero is in the table, which satisfies the disabled case.
+                    vec![(q * a, table)]
+                });
+
+                FaultyCircuitConfig { a, q, table }
+            }
+
+            fn without_witnesses(&self) -> Self {
+                Self {}
+            }
+
+            fn synthesize(
+                &self,
+                config: Self::Config,
+                mut layouter: impl Layouter<Fp>,
+            ) -> Result<(), Error> {
+                layouter.assign_table(
+                    || "Doubling table",
+                    |mut table| {
+                        (0..(1 << (K - 1)))
+                            .map(|i| {
+                                table.assign_cell(
+                                    || format!("table[{}] = {}", i, 2 * i),
+                                    config.table,
+                                    i,
+                                    || Ok(Fp::from(2 * i as u64)),
+                                )
+                            })
+                            .fold(Ok(()), |acc, res| acc.and(res))
+                    },
+                )?;
+
+                layouter.assign_region(
+                    || "Good synthesis",
+                    |mut region| {
+                        // Enable the lookup on rows 0 and 1.
+                        config.q.enable(&mut region, 0)?;
+                        config.q.enable(&mut region, 1)?;
+
+                        // Assign a = 2 and a = 6.
+                        region.assign_advice(|| "a = 2", config.a, 0, || Ok(Fp::from(2)))?;
+                        region.assign_advice(|| "a = 6", config.a, 1, || Ok(Fp::from(6)))?;
+
+                        Ok(())
+                    },
+                )?;
+
+                layouter.assign_region(
+                    || "Faulty synthesis",
+                    |mut region| {
+                        // Enable the lookup on rows 0 and 1.
+                        config.q.enable(&mut region, 0)?;
+                        config.q.enable(&mut region, 1)?;
+
+                        // Assign a = 4.
+                        region.assign_advice(|| "a = 4", config.a, 0, || Ok(Fp::from(4)))?;
+
+                        // BUG: Assign a = 5, which doesn't exist in the table!
+                        region.assign_advice(|| "a = 5", config.a, 1, || Ok(Fp::from(5)))?;
+
+                        Ok(())
+                    },
+                )
+            }
+        }
+
+        let prover = MockProver::run(K, &FaultyCircuit {}, vec![]).unwrap();
+        assert_eq!(
+            prover.verify(),
+            Err(vec![VerifyFailure::Lookup {
+                lookup_index: 0,
+                location: LookupFailure::InRegion {
+                    region: (2, "Faulty synthesis".to_owned()).into(),
+                    offset: 1,
+                }
             }])
         );
     }