+++ src/plonk/prover.rs
@@ -43,14 +43,12 @@ pub fn create_proof<
 ) -> Result<(), Error> {
     for instance in instances.iter() {
         if instance.len() != pk.vk.cs.num_instance_columns {
-            return Err(Error::IncompatibleParams);
+            return Err(Error::InvalidInstances);
         }
     }
 
     // Hash verification key into transcript
-    pk.vk
-        .hash_into(transcript)
-        .map_err(|_| Error::TranscriptError)?;
+    pk.vk.hash_into(transcript)?;
 
     let domain = &pk.vk.domain;
     let mut meta = ConstraintSystem::default();
@@ -94,9 +92,7 @@ pub fn create_proof<
             drop(instance_commitments_projective);
 
             for commitment in &instance_commitments {
-                transcript
-                    .common_point(*commitment)
-                    .map_err(|_| Error::TranscriptError)?;
+                transcript.common_point(*commitment)?;
             }
 
             let instance_polys: Vec<_> = instance_values
@@ -132,6 +128,7 @@ pub fn create_proof<
         .zip(instances.iter())
         .map(|(circuit, instances)| -> Result<AdviceSingle<C>, Error> {
             struct WitnessCollection<'a, F: Field> {
+                k: u32,
                 pub advice: Vec<Polynomial<Assigned<F>, LagrangeCoeff>>,
                 instances: &'a [&'a [F]],
                 usable_rows: RangeTo<usize>,
@@ -172,7 +169,7 @@ pub fn create_proof<
                     row: usize,
                 ) -> Result<Option<F>, Error> {
                     if !self.usable_rows.contains(&row) {
-                        return Err(Error::BoundsFailure);
+                        return Err(Error::not_enough_rows_available(self.k));
                     }
 
                     self.instances
@@ -196,7 +193,7 @@ pub fn create_proof<
                     AR: Into<String>,
                 {
                     if !self.usable_rows.contains(&row) {
-                        return Err(Error::BoundsFailure);
+                        return Err(Error::not_enough_rows_available(self.k));
                     }
 
                     *self
@@ -263,6 +260,7 @@ pub fn create_proof<
             let unusable_rows_start = params.n as usize - (meta.blinding_factors() + 1);
 
             let mut witness = WitnessCollection {
+                k: params.k,
                 advice: vec![domain.empty_lagrange_assigned(); meta.num_advice_columns],
                 instances,
                 // The prover will not be allowed to assign values to advice
@@ -303,9 +301,7 @@ pub fn create_proof<
             drop(advice_commitments_projective);
 
             for commitment in &advice_commitments {
-                transcript
-                    .write_point(*commitment)
-                    .map_err(|_| Error::TranscriptError)?;
+                transcript.write_point(*commitment)?;
             }
 
             let advice_polys: Vec<_> = advice
@@ -498,9 +494,7 @@ pub fn create_proof<
 
         // Hash each instance column evaluation
         for eval in instance_evals.iter() {
-            transcript
-                .write_scalar(*eval)
-                .map_err(|_| Error::TranscriptError)?;
+            transcript.write_scalar(*eval)?;
         }
     }
 
@@ -520,9 +514,7 @@ pub fn create_proof<
 
         // Hash each advice column evaluation
         for eval in advice_evals.iter() {
-            transcript
-                .write_scalar(*eval)
-                .map_err(|_| Error::TranscriptError)?;
+            transcript.write_scalar(*eval)?;
         }
     }
 
@@ -537,9 +529,7 @@ pub fn create_proof<
 
     // Hash each fixed column evaluation
     for eval in fixed_evals.iter() {
-        transcript
-            .write_scalar(*eval)
-            .map_err(|_| Error::TranscriptError)?;
+        transcript.write_scalar(*eval)?;
     }
 
     let vanishing = vanishing.evaluate(x, xn, domain, transcript)?;
@@ -611,5 +601,5 @@ pub fn create_proof<
         // We query the h(X) polynomial at x
         .chain(vanishing.open(x));
 
-    multiopen::create_proof(params, transcript, instances).map_err(|_| Error::OpeningError)
+    multiopen::create_proof(params, transcript, instances).map_err(|_| Error::Opening)
 }